+++ app/soapbox/components/ui/index.ts
@@ -41,6 +41,7 @@ export { default as PhoneInput } from './phone-input/phone-input';
 export { default as ProgressBar } from './progress-bar/progress-bar';
 export { default as RadioButton } from './radio-button/radio-button';
 export { default as Select } from './select/select';
+export { default as Slider } from './slider/slider';
 export { default as Spinner } from './spinner/spinner';
 export { default as Stack } from './stack/stack';
 export { default as Streamfield } from './streamfield/streamfield';